+++ CHANGELOG.md
@@ -35,6 +35,10 @@
   - Improved menubar blur
 - Add Nvidia Kepler GOP Driver injection
   - Primarily for GPUs lacking GOPs and can't have a newer VBIOS flashed
+- Resolve Rapid Security Response support for Haswell Macs requiring KDKs
+  - Implemented via:
+    - Userspace: RSRRepair at `/etc/rc.server` (2b1c9e3)
+    - Kernelspace: RSRHelper.kext (cbe1be9)
 - Increment Binaries:
   - OpenCorePkg 0.8.8 - release
   - PatcherSupportPkg 0.8.0 - release